Make --verify-ssl choices explicit
This adds true/false as explicit choices so that the help output is a
bit friendlier.

    [--verify-ssl {true, false, yes, no}]
    --verify-ssl {true, false, yes, no}
          Enable or disable the verification of the server SSL
          certificate

In CLIs with boolean flags it is also sometimes the case that you've a
``--no-*`` variant or that you can use ``no`` or something like that. So
having to pass ``false`` is not necessarily intuitive.