Improve visual representation of SQL statements
This commit changes the lexer for tokenizing the user input from the
generic SqlLexer to the more specific PostgresLexer.

Additionally the color scheme has been changed from a custom on to
"terminal style", meaning that it will pick up the colors of the
terminal emulator (http://pygments.org/docs/styles/#ansiterminalstyle).