Added `beep` command to toggle console beeping upon SQL command completion
I've found that having a console beep at the conclusion of a long
running SQL query came in really useful for multitasking. I would
usually do this through the pattern of:

```
crash -c "select * from sys.summits;" && echo -e "\a"
```

This PR adds a command that lets crash do this instead, by sending the
bell character to the console.