Rollback release workflow changes
The `shiv` bundling step is flaky because the package on PyPI is not
immediately available (and it was also broken because it missed the
`upload_url` property).

This rolls back the changes to do the bundling via Jenkins job. Then the
bundle is also on our CDN in the same location as before. So we don't
hvae to change other components (like the docker builds, which fetch the
bundle from the CDN).