Fix ctrl+r search cancelling not returning focus
When in vi-mode, pressing ctrl+r starts the search mode. We would expect
that pressing ctrl+c to close the search mode would return focus back to
the text input, but this never happened.

What was happening is that, when you press ctrl-r, a new control element
is created called a SearchBufferControl, which is then focused on.
Pressing esc closes this control, and returns focus to the previous
control (the one created by our default text buffer).

For some reason, the keyboard interrupt exception isn't caught by this
control, and is instead bubbled up to our application. Consequently,
this SearchBufferControl remains in focus and cannot be escaped.

This PR sets it so that when a KeyboardInterrupt is caught and the
current buffer control is a search buffer control, focus is instead
given back to the previous control. This should address #320

I'm not sure if this is a bug with the prompt toolkit or not. Will look
into it further.